refactor(app): move the skill manager screen state into ainb-app
SkillsScreenData, the browse, library, source-preview and input view states,
the discovery banner logic and the selection reducers move to
ainb_app::components::skill_manager_screen. ainb-app gains the
ainb-skill-core and ainb-cli dependencies they use; neither pulls in
ratatui or crossterm. The screen renderer and its tests stay in ainb-core.
