+++ ainb-tui/crates/ainb-app/src/components/mod.rs
@@ -5,4 +5,5 @@
 pub mod config_popup;
 pub mod onboarding;
 pub mod setup_menu;
+pub mod skill_manager_screen;
 pub mod skills;
